+++ lib/active_record/connection_adapters/sqlserver/type/smalldatetime.rb
@@ -19,7 +19,7 @@ module ActiveRecord
           end
 
           def apply_seconds_precision(value)
-            value.change usec: 0
+            value.change usec: 0 if value
           end
 
         end