+++ test/cases/schema_dumper_test_sqlserver.rb
@@ -85,6 +85,7 @@ class SchemaDumperTestSQLServer < ActiveRecord::TestCase
     # Our type methods.
     columns['real_col'].sql_type.must_equal         'real'
     columns['money_col'].sql_type.must_equal        'money'
+    columns['smalldatetime_col'].sql_type.must_equal 'smalldatetime'
     columns['datetime2_col'].sql_type.must_equal    'datetime2(7)'
     columns['datetimeoffset'].sql_type.must_equal   'datetimeoffset(7)'
     columns['smallmoney_col'].sql_type.must_equal   'smallmoney'
@@ -100,6 +101,7 @@ class SchemaDumperTestSQLServer < ActiveRecord::TestCase
     columns['json_col'].sql_type.must_equal         'nvarchar(max)'
     assert_line :real_col,          type: 'real',           limit: nil,           precision: nil,   scale: nil,  default: nil
     assert_line :money_col,         type: 'money',          limit: nil,           precision: 19,    scale: 4,    default: nil
+    assert_line :smalldatetime_col, type: 'smalldatetime',  limit: nil,           precision: nil,   scale: nil,  default: nil
     assert_line :datetime2_col,     type: 'datetime',       limit: nil,           precision: 7,     scale: nil,  default: nil
     assert_line :datetimeoffset,    type: 'datetimeoffset', limit: nil,           precision: 7,     scale: nil,  default: nil
     assert_line :smallmoney_col,    type: 'smallmoney',     limit: nil,           precision: 10,    scale: 4,    default: nil